requirements.txt
Django==4.1
django_countries==7.3.2
django_rest_knox==4.2.0
django-rest-swagger==2.2.0
django_rest_passwordreset==1.2.1
djangorestframework==3.13.1
drf_yasg==1.21.3
knox==0.1.14
Django==4.1
drf-yasg==1.21.3
openapi-codec==1.3.2
Pillow==9.2.0
coverage==6.4.4
pytest==7.1.2
pytest-cov==3.0.0
dynaconf==3.1.9
factory-boy==3.2.1
Faker==12.0.1
pluggy==1.0.0
py==1.11.0
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.2.1
redgreenunittest==0.1.1
validators==0.20.0
